Refactor: Extract new mod `binstall/resolve.rs`
